Added a Microsoft Visual Studio test reporter.
This is essentially the same as the basic reporter, however it formats the
output slightly differently and adds "error:" and "warning:" in the appropriate
places to have those lines show up in Visual Studio's Error List window.  This
is useful when setting the test executable to automatically run as a post-build
event.

See:
http://blogs.msdn.com/b/msbuild/archive/2006/11/03/msbuild-visual-studio-aware-error-messages-and-message-formats.aspx