restrict nullptr detection to gcc 4 (closes #526)
The old detection logic here assumed that all versions of gcc >= 5 would
provide nullptr and `std::nullptr_t`; this is only true in gcc >= 5 when
the standard is explicitly set (e.g. `std=c++11`), and other detection
logic handles that.

This change allows Catch-using programs to compile with gcc >= 5, when
using the default (C++98) standard.