add no-rtti and no-exception support to catch2
(This is a work-in-progress)

try/catch/throw/typeid become illegal keywords when their compiler
support is disabled.

By quering the compiler __has__feature we can determine if those
keywords would be legal to use or not.

For RTTI, disable the offending tests and replace dynamic_cast with
virtual calls.

For exceptions, replace try/catch with macros and throw with an
std::terminate.

Limitations: Currently when a REQUIRE() call fails it immediately
calls std::terminate, causing the rest of the test run to abort.

Correctness: Running the SelfTest main by itself passes 100%.

However running approvalTests script (without rtti and exceptions) fails,
largely because the .txt files that are diffed against have test output
that is now disabled without exceptions.