Amended previous commit
Ah, MSVC, my old nemesis, we meet again. This commit amends the
previous commit to re-introduce the following using declarations inside
the CustomAllocator class:

- using std::allocator<T>::address;
- using std::allocator<T>::construct;
- using std::allocator<T>::max_size;
- using std::allocator<T>::destroy;

When building with MSVC and using a C++ standard older than
C++17. These were re-introduced because MSVC throws an error if they're
not present when compiling with C++11 or C++14.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>