ObjectStorage: fix UB (strict aliasing violation)
ObjectStorage is acting like a std::optional, manually managing the
lifetime of an object placed into a char array buffer. C++17 changes
the lifetime rules around such patterns, with the effect that this
code now exhibits undefined behaviour, because &data is a pointer to a
struct containing a char array, not a pointer to T, and it remains
that way even after the static_cast.

The fix is to use std::launder(), which, for C++14 compilers, we hide
behind a macro.

The similar Option type is not affected, since it has an additional T*
member that stores the result of placement new, and therefore doesn't
alias anything.

Fixes #2638.